Add translation units for inputarray, maxcolarray & minrowarray functions
