+++ namestnikov.kirill/P4/input_array.hpp
@@ -1,0 +1,7 @@
+#ifndef INPUTARRAY_H
+#define INPUTARRAY_H
+#include <cstddef>
+#include <iostream>
+#include <fstream>
+size_t inputArray(std::istream & in, int * a, size_t s, size_t toread);
+#endif
