+++ namestnikov.kirill/P4/max_col_array.hpp
@@ -1,0 +1,5 @@
+#ifndef MAXCOLARRAY_H
+#define MAXCOLARRAY_H
+#include <cstddef>
+int * maxColArray(size_t result, size_t cols, int * matrix);
+#endif
