+++ namestnikov.kirill/P4/min_row_array.hpp
@@ -1,0 +1,5 @@
+#ifndef MINROWARRAY_H
+#define MINROWARRAY_H
+#include <cstddef>
+int * minRowArray(size_t result, size_t cols, int * matrix);
+#endif
